Correct install information for gnupg2 on OS X
The homebrew package name given in the readme was for the wrong version of gpg. It now has the correct package for gnupg2, and also a link to the homesite (as this is a default part of the OS) and to the full GPGTools suite, which is the full set of GPG tools for OS X (including gnupg2).